backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.7.0
httpx>=0.27.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0

# LLM Integration (Phase 3)
langchain>=0.3.0
langchain-google-genai>=2.0.0
google-generativeai>=0.8.0

# Phase 7: Hardening & QA
slowapi>=0.1.9
tenacity>=8.2.0
cachetools>=5.3.0
